services/order-orchestrator/pyproject.toml
[tool.poetry]
name = "order-orchestrator"
version = "0.1.0"
description = "Order batching, validation, routing, and lifecycle tracking"
authors = ["Your Team <dev@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
alembic = "^1.12.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
aiokafka = "^0.10.0"
prometheus-client = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
aiosqlite = "^0.19.0"
pytest-httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
